refactor: code review #52 - fix floating-point money precision by replacing double with int64_t for all monetary amounts in protocol and server code (eliminates precision errors in financial calculations, uses cents as base unit for exact arithmetic), remove duplicate cmake_minimum_required from cmake/Sanitizers.cmake (redundant requirement already present in root CMakeLists.txt), simplify websocket_session destructor by removing unnecessary try-catch blocks (noexcept declaration already prohibits exceptions, cleanup operations use error_code pattern that doesn't throw, race condition mitigation achieved through proper RAII and error_code usage)

+++ src/server/config.hpp
@@ -32,7 +32,7 @@ struct config {
     static constexpr auto MAX_BACKOFF = std::chrono::seconds{30};
     
     static constexpr int PLACEHOLDER_SEAT = -1;
-    static constexpr double PLACEHOLDER_STACK = 0.0;
+    static constexpr int64_t PLACEHOLDER_STACK = 0;
     
     // WebSocket stream idle/read timeouts.
     // Set to 24 hours so the application-level timeouts (HANDSHAKE_TIMEOUT,
